Fit a fibre

fibre() fits a model.

Usage

fibre(x, ...)

# S3 method for default
fibre(x, ...)

# S3 method for data.frame
fibre(
  x,
  y,
  intercept = TRUE,
  engine = c("inla", "glmnet", "mgcv"),
  engine_options = list(),
  ...
)

# S3 method for matrix
fibre(
  x,
  y,
  intercept = TRUE,
  engine = c("inla", "glmnet", "mgcv"),
  engine_options = list(),
  ...
)

# S3 method for formula
fibre(
  formula,
  data,
  intercept = TRUE,
  family = "gaussian",
  engine = c("inla", "glmnet", "mgcv"),
  engine_options = list(),
  ...
)

# S3 method for recipe
fibre(
  x,
  data,
  intercept = TRUE,
  engine = c("inla", "glmnet", "mgcv"),
  engine_options = list(),
  ...
)

Arguments

x

Depending on the context:

A data frame of predictors.

A matrix of predictors.

A recipe specifying a set of preprocessing steps created from recipes::recipe().

...

Not currently used, but required for extensibility.

y

When x is a data frame or matrix, y is the outcome specified as:

A data frame with 1 numeric column.

A matrix with 1 numeric column.

A numeric vector.

formula

A formula specifying the outcome terms on the left-hand side, and the predictor terms on the right-hand side.

data

When a recipe or formula is used, data is specified as:

A data frame containing both the predictors and the outcome.

Value

A fibre object.

Examples

predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
#mod <- fibre(predictors, outcome)

# Formula interface
#mod2 <- fibre(mpg ~ ., mtcars)

# Recipes interface
#library(recipes)
#rec <- recipe(mpg ~ ., mtcars)
#rec <- step_log(rec, disp)
#mod3 <- fibre(rec, mtcars)