Tidy Model Results

Tidy Model Results

Usage

# S3 method for fibre
tidy(
  x,
  effects = c("fixed", "rates", "random", "hyper"),
  conf.type = c("cred.int", "marginals"),
  indexes = NULL,
  ...
)

Arguments

x

A fibre model object

effects

Which effects do you want tidied? One of: "fixed", for fixed effects, "random" for random effects, or "hyper" for the hyper-parameters of the random effects. Can also be "rates", which is a synonym for "random", since the random effects are rates of trait evolution along phylogenetic edges.

conf.type

What kind of confidence interval. Choices are: "cred.int" for approximate Bayesian marginal credible intervals. or "marginals" for the full approximate marginal distributions, as a data.frame with value and y.value columns. value is the value of the parameter, and y.value is the marginal posterior density (e.g. what value is the x axis and y.value is the y axis when plotting the posterior density).

indexes

If effects = "random" or effects = "rates", this is a vector of indices to retrieve particular random effects. Default is to return all random effects, however, this can be slow for retrieving the marginals.

...

Not used.

Value

A tidy tibble with information about the fitted model parameters.